Full refactor into packages,import statement cleanup, commented all my code, broke HeliocentricActivity into two classes (one for calculations)
