+++ app/src/main/java/ca/mun/engi5895/stargazer/DataInstallation/ErrorDialogFragment.java
@@ -1,15 +1,10 @@
-package ca.mun.engi5895.stargazer;
+package ca.mun.engi5895.stargazer.DataInstallation;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.app.DialogFragment;
 import android.content.DialogInterface;
 import android.os.Bundle;
-
-import android.content.SharedPreferences;
-import android.content.res.AssetManager;
-import android.os.Environment;
-import android.preference.PreferenceManager;
 
 import ca.mun.engi5895.stargazer.R;
