+++ app/src/main/java/ca/mun/engi5895/stargazer/DataInstallation/celestrakData.java
@@ -1,4 +1,4 @@
-package ca.mun.engi5895.stargazer;
+package ca.mun.engi5895.stargazer.DataInstallation;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,10 +6,8 @@ import android.content.Intent;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
