+++ app/src/main/java/ca/mun/engi5895/stargazer/DataInstallation/saveFileIntent.java
@@ -1,0 +1,192 @@
+package ca.mun.engi5895.stargazer.DataInstallation;
+
+import android.app.IntentService;
+import android.content.Context;
+import android.content.Intent;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.URL;
+
+/**
+ * Created by john on 2018-02-24.
+ */
+
+public class saveFileIntent extends IntentService {
+
+    //These are like parameters. See SettingsActivity onClick() method to see how they are used
+    public static final String URL = "urlpath";
+    public static final String FILENAME = "filename";
+    public static final String FILEPATH = "filepath";
+
+
+    public saveFileIntent() {
+        super("ca.mun.engi5895.stargazer.DataInstallation.saveFileIntent");
+    }
+
+    @Override
+    protected void onHandleIntent(Intent intent) {
+
+        String urlPath = intent.getStringExtra(URL);
+        String fileName = intent.getStringExtra(FILENAME);
+
+        File output = new File(getFilesDir(), fileName);
+
+        if (output.exists()) {
+            output.delete();
+        }
+
+        InputStream stream = null;
+        FileOutputStream fos = null;
+
+        try {
+            URL url = new URL(urlPath);
+            stream = url.openConnection().getInputStream();
+            InputStreamReader reader = new InputStreamReader(stream);
+            fos = new FileOutputStream(output.getPath());
+            int next = -1;
+            while ((next = reader.read()) != -1) {
+                fos.write(next);
+            }
+            // successfully finished
+
+            //result = Activity.RESULT_OK;
+
+
+
+            //Handler thing is to let you make a toast on a dead thread, has to be in for toast to work
+            /*Handler mHandler = new Handler(getMainLooper());
+            mHandler.post(new Runnable() {
+                @Override
+                public void run() {
+                    Toast.makeText(getApplicationContext(), "Files Downloaded", Toast.LENGTH_SHORT).show();
+                }
+            });*/
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            if (stream != null) {
+                try {
+                    stream.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+
+
+
+        String filename = "favoriteSats.txt";
+        String fileContents = "Favorite Entities List";
+        File file = new File(this.getFilesDir(), filename);
+        FileOutputStream outputStream;
+
+        try {
+            outputStream = openFileOutput(filename, Context.MODE_PRIVATE);
+            outputStream.write(fileContents.getBytes());
+            outputStream.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+
+    }
+
+/*
+    private static boolean dirChecker(String pathname) {
+
+        File theDir = new File(pathname);
+
+        boolean result = false;
+        // if the directory does not exist, create it
+        if (!theDir.exists())
+
+        {
+            System.out.println("creating directory: " + theDir.getName());
+            //result = false;
+
+            try {
+                theDir.mkdir();
+                result = true;
+            } catch (SecurityException se) {
+                //handle it
+            }
+            if (result) {
+                System.out.println("DIR created");
+            }
+        }
+        return result;
+    }
+*/
+
+
+/*
+    public static void unzip(String _zipFile, String _targetLocation) {
+
+        //create target location folder if not exist
+        dirChecker(_targetLocation);
+
+        try {
+            System.out.println(new File(".").getAbsolutePath());
+            FileInputStream fin = new FileInputStream(_zipFile);
+            ZipInputStream zin = new ZipInputStream(fin);
+            ZipEntry ze = null;
+            while ((ze = zin.getNextEntry()) != null) {
+
+                //create dir if required while unzipping
+                if (ze.isDirectory()) {
+                    dirChecker(ze.getName());
+                } else {
+                    FileOutputStream fout = new FileOutputStream(_targetLocation + ze.getName());
+                    for (int c = zin.read(); c != -1; c = zin.read()) {
+                        fout.write(c);
+                    }
+
+                    zin.closeEntry();
+                    fout.close();
+                }
+
+            }
+            zin.close();
+        } catch (Exception e) {
+           System.out.println(e);
+        }
+    }public static void unzip(String _zipFile, String _targetLocation) {
+
+        //create target location folder if not exist
+        dirChecker(_targetLocation);
+
+        try {
+            System.out.println(new File(".").getAbsolutePath());
+            FileInputStream fin = new FileInputStream(_zipFile);
+            ZipInputStream zin = new ZipInputStream(fin);
+            ZipEntry ze = null;
+            while ((ze = zin.getNextEntry()) != null) {
+
+                //create dir if required while unzipping
+                if (ze.isDirectory()) {
+                    dirChecker(ze.getName());
+                } else {
+                    FileOutputStream fout = new FileOutputStream(_targetLocation + ze.getName());
+                    for (int c = zin.read(); c != -1; c = zin.read()) {
+                        fout.write(c);
+                    }
+
+                    zin.closeEntry();
+                    fout.close();
+                }
+
+            }
+            zin.close();
+        } catch (Exception e) {
+           System.out.println(e);
+        }
+    }
+    */
+}
+
+
+
