--- app/src/main/java/ca/mun/engi5895/stargazer/OrbitGaphing.java
@@ -1,8 +1,0 @@
-package ca.mun.engi5895.stargazer;
-
-/**
- * Created by noahg on 3/18/2018.
- */
-
-public class OrbitGaphing {
-}
