+++ app/src/main/res/layout/activity_heliocentric.xml
@@ -4,9 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ca.mun.engi5895.stargazer.HeliocentricActivity">
+    tools:context="ca.mun.engi5895.stargazer.Activities.HeliocentricActivity">
 
-    <ca.mun.engi5895.stargazer.SquareLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    <ca.mun.engi5895.stargazer.AndroidAesthetics.SquareLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:ap="http://schemas.android.com/apk/res-auto"
         android:id="@+id/layout"
         android:layout_width="fill_parent"
@@ -25,7 +25,7 @@
             ap:rangeTitle="range"
             ap:title="A Simple XY Plot" />
 
-    </ca.mun.engi5895.stargazer.SquareLayout>
+    </ca.mun.engi5895.stargazer.AndroidAesthetics.SquareLayout>
 
     <RelativeLayout
         android:id="@+id/topButtons"
