+++ app/src/main/res/layout/activity_maps.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ca.mun.engi5895.stargazer.GeocentricActivity">
+    tools:context="ca.mun.engi5895.stargazer.Activities.MapsActivity">
 
     <fragment xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:map="http://schemas.android.com/apk/res-auto"
@@ -15,7 +15,7 @@
         map:layout_constraintEnd_toEndOf="parent"
         map:layout_constraintStart_toStartOf="parent"
         map:layout_constraintTop_toTopOf="parent"
-        tools:context="ca.mun.engi5895.stargazer.MapsActivity" />
+        tools:context="ca.mun.engi5895.stargazer.Activities.MapsActivity" />
 
     <TimePicker
         android:id="@+id/timePicker"
